add solution day12 part 2

+++ day12/Records.h
@@ -18,7 +18,7 @@ public:
 	Records(std::vector<SpringRow> const &rows_): rows(rows_) {};
 	virtual ~Records() {};
 
-	int get_options() const;
+	long get_options() const;
 
 	static Records from_file(std::string const &filename);
 private:
